+++ tests/corpus/raw-baseline/arm64_O2/adler32.c
@@ -16,10 +16,10 @@ uint64_t sym__adler32(uint64_t X0_0, uint64_t X1_0)
             uint8_t tmp_25400_2 = *tmp_7400_2;
             uint32_t tmp_12280_2 = (uint32_t)tmp_25400_2 + X9_1;
             tmp_2b000_2 = tmp_12280_2 - (uint32_t)((uint64_t)tmp_12280_2 * 0x80078071 >> 47) * 0xfff1;
-            X9_1 = (uint32_t)tmp_2b000_2;
+            X9_1 = tmp_2b000_2;
             uint32_t tmp_12280_3 = X8_1 + tmp_2b000_2;
             tmp_2b000_3 = tmp_12280_3 - (uint32_t)((uint64_t)tmp_12280_3 * 0x80078071 >> 47) * 0xfff1;
-            X8_1 = (uint32_t)tmp_2b000_3;
+            X8_1 = tmp_2b000_3;
             uint8_t TMPZR_4 = X1_0 == 1;
             X1_0--;
             if (TMPZR_4) {
